move credits tests to testing module
