+++ contracts/credits/src/lib.rs
@@ -1,3 +1,7 @@
 pub mod contract;
 pub mod msg;
 pub mod state;
+
+#[allow(clippy::unwrap_used)]
+#[cfg(test)]
+mod testing;
